Video #8 shows a way to use the onClick event and how to get access to the event and its properties.
for the handleclick function we can see that the function gets access automatically to the event so if we want we can add it as a function parameter and then use it.

For the handleClickAgain function it doesnt work quite the same because we are using an anonymous function to call the function we want to send with a parameter so we need to send the event "e" within the anonymous parenthesis (e) => {useEvent(e)} then we can send the event "e" to the parameters and use it. within the definition of our function.